+++ pyannote/core/scores.py
@@ -32,9 +32,9 @@ import numpy as np
 from pandas import Index, MultiIndex, DataFrame, pivot_table
 
 from . import PYANNOTE_SEGMENT, PYANNOTE_TRACK, PYANNOTE_LABEL, PYANNOTE_SCORE
-from annotation import Annotation, Unknown
-from segment import Segment
-from timeline import Timeline
+from .annotation import Annotation, Unknown
+from .segment import Segment
+from .timeline import Timeline
 
 
 class Scores(object):
@@ -197,6 +197,9 @@ class Scores(object):
         return len(self.annotation_)
 
     def __nonzero__(self):
+        return self.__bool__()
+
+    def __bool__(self):
         """False if annotation is empty"""
         return True if self.annotation_ else False
 
@@ -300,11 +303,14 @@ class Scores(object):
         # make sure segment/track pairs are sorted
         self._reindexIfNeeded()
 
+        labels = self.labels()
+
         # yield one (segment, track, label) tuple per loop
         for index, columns in self.dataframe_.iterrows():
             segment = Segment(*index[:-1])
             track = index[-1]
-            for label, value in dict(columns).iteritems():
+            for label in labels:
+                value = columns[label]
                 if not np.isnan(value):
                     yield segment, track, label, value